requirements.txt
Django==5.1.2
django-ckeditor==6.7.1
django-jazzmin==3.0.1
django-taggit==6.1.0
easy-thumbnails==2.10
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv
gunicorn
whitenoise
django-cloudinary-storage